Splitting up configuration into files
